Split the tests back into several files
