+++ cloudify_cli/tests/resources/blueprints/helloworld/blueprint.yaml
@@ -3,6 +3,14 @@ tosca_definitions_version: cloudify_dsl_1_2
 imports:
     - sample_app.yaml
 
+inputs:
+    key1:
+        default: default_val1
+    key2:
+        default: default_val2
+    key3:
+        default: default_val3
+
 node_templates:
     node1:
         type: type1
